finish migrating accounts api
