+++ client/src/hooks/mutations/accounts/usePermanentDeleteAccountMutation.ts
@@ -1,0 +1,37 @@
+import { notifications } from "@mantine/notifications";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { AxiosError } from "axios";
+import {
+  accountsQueryKey,
+  balancesQueryKey,
+  institutionsQueryKey,
+  transactionsQueryKey,
+  translateAxiosError,
+} from "~/helpers/requests";
+import { useAuth } from "~/providers/AuthProvider/AuthProvider";
+
+export const usePermanentDeleteAccountMutation = () => {
+  const queryClient = useQueryClient();
+  const { request } = useAuth();
+
+  return useMutation({
+    mutationFn: async (accountId: string) =>
+      await request({
+        url: "/api/account/permanentDelete",
+        method: "DELETE",
+        params: { guid: accountId },
+      }),
+    onSuccess: async () => {
+      await queryClient.invalidateQueries({ queryKey: [institutionsQueryKey] });
+      await queryClient.invalidateQueries({ queryKey: [accountsQueryKey] });
+      await queryClient.invalidateQueries({ queryKey: [transactionsQueryKey] });
+      await queryClient.invalidateQueries({ queryKey: [balancesQueryKey] });
+    },
+    onError: (error: AxiosError) => {
+      notifications.show({
+        color: "var(--button-color-destructive)",
+        message: translateAxiosError(error),
+      });
+    },
+  });
+};
